DEVEXP-512: Refactor ClientParameter builders
Also fixed a small bug with the TDE template validation.
